now it takes all images in a certain dir and make a timelapse

+++ pylapse.py
@@ -1,16 +1,21 @@
 import cv2
+import os
 
-img1 = cv2.imread('appFolder/static/photos/1.jpg')
-img2 = cv2.imread('appFolder/static/photos/2.jpg')
-img3 = cv2.imread('appFolder/static/photos/3.jpg')
+directory = './test/'
+l = []
+time_one_im = 0.2
+fps = 1/time_one_im # may need a cast to int
 
-height , width , layers =  img1.shape
-
-video = cv2.VideoWriter('video.avi',cv2.VideoWriter_fourcc(*"MJPG"), 10,(width,height))
-
-video.write(img1)
-video.write(img2)
-video.write(img3)
-
-cv2.destroyAllWindows()
-video.release()
+if os.path.exists(path=directory):
+    l = os.listdir(directory)
+    l = sorted(l)
+    height , width , layers =  cv2.imread(directory + l[0]).shape
+    video = cv2.VideoWriter('video.avi', cv2.VideoWriter_fourcc(*"MJPG"),
+                            fps, (width, height))
+    for img_path in l:
+        video.write(cv2.imread(directory + img_path))
+    cv2.destroyAllWindows()
+    video.release()
+    print('video done')
+else:
+    print('cannot find the directory - ' + directory)
